Refactoring issue,kanban, milestone socket and change their types(Close #28)

+++ server/src/module/data/kanbanManager.ts
@@ -4,6 +4,7 @@ import { getJsonData, isExists, setJsonData } from "./fileManager";
 import fs from "fs";
 import { v4 as uuidv4 } from "uuid";
 import log from "../log";
+import DataProjectManager from "./projectManager";
 
 export default class DataKanbanManager {
     static isExists(kanbanUUID: string) {
@@ -58,29 +59,41 @@ export default class DataKanbanManager {
             fs.mkdirSync(this.getKanbanPath(), { recursive: true });
         }
 
-        const kanbanList: TkanbanData[] = [];
-        fs.readdirSync(this.getKanbanPath())
+        return fs
+            .readdirSync(this.getKanbanPath())
             .filter((kanban) => {
                 const kanbanData = kanban !== "milestoneListInfo.json" ? this.getKanbanInfo(kanban) : undefined;
                 return (
                     kanban !== "milestoneListInfo.json" &&
                     (options.uuid === undefined || options.uuid === kanbanData?.uuid) &&
                     (options.column === undefined || (options.column as string) in (kanbanData?.columns as string[])) &&
-                    (options.projectName === undefined || options.projectName === kanbanData?.projectName) &&
-                    (options.milestone === undefined || options.milestone === kanbanData?.milestone)
+                    (options.projectName === undefined || options.projectName === kanbanData?.projectName)
                 );
             })
-            .map((kanban) => {
+            .reduce((kanbanList: TkanbanData[], kanban: string) => {
                 kanbanList.push(this.getKanbanInfo(kanban) as TkanbanData);
-            });
-        return kanbanList;
+                return kanbanList;
+            }, []);
     }
 
     static create(kanbanData: TkanbanCreateData) {
         const kanbanUUID = uuidv4();
         fs.mkdirSync(this.getKanbanPath(kanbanUUID), { recursive: true });
 
-        if (!this.setKanbanInfo(kanbanUUID, { uuid: kanbanUUID, columns: ["backlog", "todo", "in progress", "Done"], totalIssue: 0, doneIssue: 0, ...kanbanData })) {
+        if (
+            !fs
+                .readdirSync(DataProjectManager.getProjectDefaultPath())
+                .reduce((projectIdList: string[], projectId: string) => {
+                    projectIdList.push(DataProjectManager.getProjectInfo(projectId)?.projectName as string);
+                    return projectIdList;
+                }, [])
+                .includes(kanbanData.projectName)
+        ) {
+            log.error(`[DataKanbanManager] create -> could not find projectName`);
+            return undefined;
+        }
+
+        if (!this.setKanbanInfo(kanbanUUID, { uuid: kanbanUUID, columns: ["backlog", "todo", "in progress", "Done"], totalIssue: 0, doneIssue: 0, nextIssue: 0, ...kanbanData })) {
             log.error(`[DataKanbanManager] create -> fail to setKanbanInfo`);
             return undefined;
         }
@@ -101,7 +114,12 @@ export default class DataKanbanManager {
         log.info(`kanbandata updated: ${JSON.stringify({ ...this.getKanbanInfo(kanbanUUID), ...kanbanData })}`);
         return true;
     }
+
     static delete(kanbanUUID: string) {
+        if (kanbanUUID === undefined || !fs.readdirSync(this.getKanbanPath()).includes(kanbanUUID)) {
+            log.error(`[DataKanbanManager] delete -> kanban uuid is not in kanbanList`);
+            return false;
+        }
         fs.rmdirSync(this.getKanbanPath(kanbanUUID), { recursive: true });
         return true;
     }
